rename RenativeConfigFile to ConfigFileBuildConfig
