+++ packages/template-starter/package.json
@@ -128,5 +128,11 @@
             "last 1 firefox version",
             "last 1 safari version"
         ]
+    },
+    "dependencies": {
+        "@rnv/engine-rn": "1.0.0-canary.7",
+        "@rnv/engine-rn-tvos": "1.0.0-canary.7",
+        "@rnv/engine-rn-next": "1.0.0-canary.7",
+        "@rnv/engine-rn-electron": "1.0.0-canary.7"
     }
 }
